requirements.txt
numpy
python-opencv
matplotlib
scipy
pillow